meson: Add fallback discovery for libxml2 and sqlite3 on MSVC
The build system for Visual Studio for libxml2, as well as the typical
amalgamation sources for sqlite3 that is used for Visual Studio builds
do not generate pkg-config files.  So, on Visual Studio builds, we need
to manually look for these dependencies when their pkg-config files
cannot be found.

This is done because hand-crafting pkg-config files is a tedious process
that can be error-prone.