meson: Improve check for MIT Kerberos on Visual Studio
Visual Studio builds will not be able to run the krb5-config script,
even if it is available, and it is not generated with the MIT Kerberos
build system for Visual Studio.

So, we need to manually check whether we have the gssapi/gssapi.h
header, and whether gssapi32.lib (32-build builds) or gssapi64.lib
(x64 builds) is available, to properly enable MIT Kerberos support on
Visual Studio builds.