SoupSessionFeature: Keep features alive until all messages have been unqueued
Features can be removed at any point from a session, and if they are
removed while there are messages in the queue, it is possible for
callbacks connected to the message signals to be called after the feature
has been disposed. Adding a reference in request_queue()
and removing it in request_enqueue() ensures that the feature will not
be disposed too early.

Add a test case for the SoupCookieJar feature that reproduces this crash.

Fixes #130