Fix xml2 and sqlite fallbacks
We should try find_library() before fallback to subproject, for
platforms that does not ship the .pc file.

Also the return value of cc.has_header() is currently not used, those
statements have no effect. Fix this by using has_headers kwarg of
find_library(), a new syntax from Meson 0.50.0. We want to find the
library AND its headers, or fallback to subproject.