hsts: accept the Strict-Transport-Security header regardless of casing
For some reason a strcmp() was used where a case-insensitive search should
be used. This was further hidden by the fact that SoupMessageHeaders interns
the name of headers, so the first time a header is seen by the library will
be the casing that is used, rendering the case-insensitivity test useless.

Fix the string comparison and move the test for sensitivity to the top so that
it runs first and the interned STS header name is the allcaps one that will
allow the test to work.